# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quablog API - Q&A and Blogging Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  displayName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export type _UserCount = {
  questions?: number;
  answers?: number;
  blogs?: number;
} | null;

export interface User {
  id: number;
  username: string;
  email: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  role: UserRole;
  isBlocked: boolean;
  createdAt: string;
  _count?: _UserCount;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  admin: "admin",
} as const;

export type _UserProfileCount = {
  questions?: number;
  answers?: number;
  blogs?: number;
};

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  role: UserProfileRole;
  createdAt: string;
  _count: _UserProfileCount;
}

export interface UsersListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface UpdateUserRequest {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
}

export interface BlockUserRequest {
  blocked: boolean;
}

export interface Question {
  id: number;
  title: string;
  body: string;
  category: string;
  tags: string[];
  votes: number;
  viewCount: number;
  answerCount: number;
  hasBestAnswer: boolean;
  authorId: number;
  author: UserProfile;
  userVote?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Answer {
  id: number;
  body: string;
  votes: number;
  isBest: boolean;
  questionId: number;
  authorId: number;
  author: UserProfile;
  userVote?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface QuestionDetail {
  id: number;
  title: string;
  body: string;
  category: string;
  tags: string[];
  votes: number;
  viewCount: number;
  hasBestAnswer: boolean;
  authorId: number;
  author: UserProfile;
  answers: Answer[];
  userVote?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface QuestionsListResponse {
  questions: Question[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateQuestionRequest {
  /**
   * @minLength 10
   * @maxLength 300
   */
  title: string;
  /** @minLength 20 */
  body: string;
  category: string;
  /** @maxItems 5 */
  tags?: string[];
}

export interface UpdateQuestionRequest {
  title?: string;
  body?: string;
  category?: string;
  tags?: string[];
}

export type VoteRequestVote =
  (typeof VoteRequestVote)[keyof typeof VoteRequestVote];

export const VoteRequestVote = {
  NUMBER_1: 1,
  NUMBER_MINUS_1: -1,
  NUMBER_0: 0,
} as const;

export interface VoteRequest {
  vote: VoteRequestVote;
}

export interface VoteResponse {
  votes: number;
  userVote?: number | null;
}

export interface CreateAnswerRequest {
  /** @minLength 10 */
  body: string;
}

export interface UpdateAnswerRequest {
  /** @minLength 10 */
  body?: string;
}

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus];

export const BlogStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface Blog {
  id: number;
  title: string;
  slug: string;
  excerpt?: string | null;
  content: string;
  coverImageUrl?: string | null;
  category: string;
  tags: string[];
  status: BlogStatus;
  readTime?: number | null;
  authorId: number;
  author: UserProfile;
  createdAt: string;
  updatedAt: string;
}

export interface BlogsListResponse {
  blogs: Blog[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateBlogRequestStatus =
  (typeof CreateBlogRequestStatus)[keyof typeof CreateBlogRequestStatus];

export const CreateBlogRequestStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface CreateBlogRequest {
  /**
   * @minLength 5
   * @maxLength 300
   */
  title: string;
  /** @minLength 50 */
  content: string;
  excerpt?: string;
  coverImageUrl?: string | null;
  category: string;
  tags?: string[];
  status?: CreateBlogRequestStatus;
}

export type UpdateBlogRequestStatus =
  (typeof UpdateBlogRequestStatus)[keyof typeof UpdateBlogRequestStatus];

export const UpdateBlogRequestStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface UpdateBlogRequest {
  title?: string;
  content?: string;
  excerpt?: string;
  coverImageUrl?: string | null;
  category?: string;
  tags?: string[];
  status?: UpdateBlogRequestStatus;
}

export interface UploadResponse {
  url: string;
  filename: string;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  tagline?: string | null;
  logoUrl?: string | null;
  faviconUrl?: string | null;
  updatedAt: string;
}

export interface UpdateSettingsRequest {
  siteName?: string;
  tagline?: string;
  logoUrl?: string | null;
  faviconUrl?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  totalQuestions: number;
  totalAnswers: number;
  totalBlogs: number;
  newUsersToday: number;
  newQuestionsToday: number;
  newBlogsToday: number;
  recentUsers: User[];
  recentQuestions: Question[];
  recentBlogs: Blog[];
}

export interface UpdateAdminProfileRequest {
  email?: string;
  currentPassword?: string;
  /** @minLength 6 */
  newPassword?: string;
}

export type ListUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type ListQuestionsParams = {
  page?: number;
  limit?: number;
  category?: string;
  search?: string;
  sort?: ListQuestionsSort;
};

export type ListQuestionsSort =
  (typeof ListQuestionsSort)[keyof typeof ListQuestionsSort];

export const ListQuestionsSort = {
  newest: "newest",
  popular: "popular",
  unanswered: "unanswered",
} as const;

export type ListBlogsParams = {
  page?: number;
  limit?: number;
  category?: string;
  search?: string;
  status?: ListBlogsStatus;
};

export type ListBlogsStatus =
  (typeof ListBlogsStatus)[keyof typeof ListBlogsStatus];

export const ListBlogsStatus = {
  published: "published",
  draft: "draft",
} as const;

export type UploadImageBodyType =
  (typeof UploadImageBodyType)[keyof typeof UploadImageBodyType];

export const UploadImageBodyType = {
  avatar: "avatar",
  blog: "blog",
  logo: "logo",
  favicon: "favicon",
} as const;

export type UploadImageBody = {
  file: Blob;
  type: UploadImageBodyType;
};