# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quablog API - Q&A and Blogging Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /questions/{questionId}/answers:
    post:
      operationId: createAnswer
      tags: [answers]
      summary: Post an answer to a question
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnswerRequest"
      responses:
        "201":
          description: Created answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
  /answers/{id}:
    patch:
      operationId: updateAnswer
      tags: [answers]
      summary: Update an answer (author or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnswerRequest"
      responses:
        "200":
          description: Updated answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"

    delete:
      operationId: deleteAnswer
      tags: [answers]
      summary: Delete an answer (author or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /answers/{id}/vote:
    post:
      operationId: voteAnswer
      tags: [answers]
      summary: Upvote or downvote an answer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResponse"
  /answers/{id}/best:
    post:
      operationId: markBestAnswer
      tags: [answers]
      summary: Mark an answer as best (question author only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as best
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── BLOGS ───────────────────────────────────────────────────────────────────
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/profile:
    patch:
      operationId: updateAdminProfile
      tags: [admin]
      summary: Update admin email and/or password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

# ─── COMPONENTS ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # Auth
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
      required: [username, email, password, displayName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # Users
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            questions:
              type: integer
            answers:
              type: integer
            blogs:
              type: integer
          nullable: true
      required: [id, username, email, displayName, role, isBlocked, createdAt]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            questions:
              type: integer
            answers:
              type: integer
            blogs:
              type: integer
      required: [id, username, displayName, role, createdAt, _count]
...
    Question:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        votes:
          type: integer
        viewCount:
          type: integer
        answerCount:
          type: integer
        hasBestAnswer:
          type: boolean
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        userVote:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, body, category, tags, votes, viewCount, answerCount, hasBestAnswer, authorId, author, createdAt, updatedAt]
...
    VoteRequest:
      type: object
      properties:
        vote:
          type: integer
          enum: [1, -1, 0]
      required: [vote]
    VoteResponse:
      type: object
      properties:
        votes:
          type: integer
        userVote:
          type: integer
          nullable: true
      required: [votes]

    # Answers
    Answer:
      type: object
      properties:
        id:
          type: integer
        body:
          type: string
        votes:
          type: integer
        isBest:
          type: boolean
        questionId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        userVote:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, body, votes, isBest, questionId, authorId, author, createdAt, updatedAt]
    CreateAnswerRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 10
      required: [body]
    UpdateAnswerRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 10

    # Blogs
    Blog:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [published, draft]
        readTime:
          type: integer
          nullable: true
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, content, category, tags, status, authorId, author, createdAt, updatedAt]
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalQuestions:
          type: integer
        totalAnswers:
          type: integer
        totalBlogs:
          type: integer
        newUsersToday:
          type: integer
        newQuestionsToday:
          type: integer
        newBlogsToday:
          type: integer
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        recentQuestions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        recentBlogs:
          type: array
          items:
            $ref: "#/components/schemas/Blog"
      required: [totalUsers, totalQuestions, totalAnswers, totalBlogs, newUsersToday, newQuestionsToday, newBlogsToday, recentUsers, recentQuestions, recentBlogs]
    UpdateAdminProfileRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6