# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quablog API - Q&A and Blogging Platform
 * OpenAPI spec version: 0.1.0
 */
import type { _UserCount } from "./_userCount";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  email: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  role: UserRole;
  isBlocked: boolean;
  createdAt: Date;
  _count?: _UserCount;
}